Moved Debian specific tests to own test file
